Interface: Do not change caption of start button
The start (and pause) button is a checkable button to support both actions.
Previously, the caption of the button was also dynamic, changing from
`Start` when unpressed to `Pause` when pressed.

However, this was really quaint behaviour and not idiomatic. The behaviour
now mimics the controls of a cassette deck where the labeling of the
button also does not change, while the behaviour on activation does.